# Don't run the benchmark once per iteration if only the first ivocation lasts longer than IterationTime